+++ src/__tests__/node-project.test.ts
@@ -1,12 +1,45 @@
 import * as yaml from 'yaml';
-import { NodeProject, NodeProjectOptions, LogLevel } from '..';
+import { NodeProject, NodeProjectOptions, LogLevel, DependenciesUpgrade } from '..';
 import { DependencyType } from '../deps';
+import { AutoUpgradeDependencies, Dependabot } from '../github';
 import * as logging from '../logging';
 import { NodePackage, NpmAccess } from '../node-package';
 import { Project } from '../project';
 import { mkdtemp, synthSnapshot, TestProject } from './util';
 
 logging.disable();
+
+describe('dependencies upgrade', () => {
+
+  test('defaults to github actions when a projen secret is defined', () => {
+
+    const project = new TestNodeProject({
+      projenSecret: 'PROJEN_SECRET',
+    });
+
+    expect(project.components.filter(c => c instanceof AutoUpgradeDependencies).length).toEqual(1);
+
+  });
+
+  test('default to dependabot when a projen secret is undefined', () => {
+
+    const project = new TestNodeProject({});
+    expect(project.components.filter(c => c instanceof Dependabot).length).toEqual(1);
+
+  });
+
+  test('can be disabled', () => {
+
+    const project = new TestNodeProject({
+      dependenciesUpgrade: DependenciesUpgrade.DISABLED,
+    });
+
+    expect(project.components.filter(c => c instanceof Dependabot).length).toEqual(0);
+    expect(project.components.filter(c => c instanceof AutoUpgradeDependencies).length).toEqual(0);
+
+  });
+
+});
 
 test('license file is added by default', () => {
   // WHEN
