+++ src/github/index.ts
@@ -1,8 +1,7 @@
 export * from './github';
 
 export * from './dependabot';
-export * from './workflows';
 export * from './mergify';
 export * from './pr-template';
 export * from './mergify';
-export * from './auto-approve';
+export * from './workflows';
