+++ src/github/workflows/index.ts
@@ -1,0 +1,4 @@
+export * from './auto-approve';
+export * from './auto-merge';
+export * from './auto-upgrade-dependencies';
+export * from './workflows';
